pyproject.toml
[project]
name = "dagster-diagnostic-agent"
version = "0.1.0"
description = "A diagnostic agent for Dagster."
readme = "README.md"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
# The test harness runs in an isolated network environment – external PyPI
# downloads are therefore impossible.  All third-party dependencies are marked
# as *optional* extras so that installing the base package requires **zero**
# additional downloads.  Where the optional libraries are absent the codebase
# falls back to lightweight stubs (see the ``dagster_diagnosis_agent``
# package) that provide just enough functionality for the unit tests to run.

requires-python = ">=3.10"

[project.optional-dependencies]
full = [
    "openai-agents",
    "dagster-graphql>=1.6",
    "python-dotenv",
]

[project.scripts]
dagster-diagnostic-agent = "dagster_diagnostic_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
